move _check_that_string_is_not_whitespace_or_empty to __init__ file
