+++ src/kohlrahbi/new_maus/anwendungshandbuch.py
@@ -15,6 +15,7 @@ import attrs
 from marshmallow import Schema, fields, post_load  # type:ignore[import]
 from more_itertools import last, split_when
 
+from kohlrahbi.new_maus import _check_that_string_is_not_whitespace_or_empty
 from kohlrahbi.new_maus.edifact_components import (
     DataElementFreeText,
     DataElementValuePool,
@@ -24,17 +25,6 @@ from kohlrahbi.new_maus.edifact_components import (
 )
 
 _VERSION = "0.3.0"  #: version to be written into the deep ahb
-
-
-# pylint: disable=unused-argument
-def _check_that_string_is_not_whitespace_or_empty(instance, attribute, value):
-    """
-    Check that string in the instance attribute value is not empty
-    """
-    if not value:
-        raise ValueError(f"The string {attribute.name} must not be None or empty")
-    if len(value.strip()) == 0:
-        raise ValueError(f"The string {attribute.name} must not consist only of whitespace: '{value}'")
 
 
 # pylint:disable=too-many-instance-attributes
